pyproject.toml
[project]
name = "easy-client"
version = "0.1.0"
description = ""
authors = [
    {name = "Bucher David",email = "45233361+Ematrion@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2 (>=3.1.6,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pydantic[email] (>=2.12.2,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "networkx (>=3.5,<4.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
]

[project.scripts]
easy-client = "easy_client.cli:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
matplotlib = "^3.10.7"

